added vercel server link

+++ src/Pages/AllSellers/AllSellers.js
@@ -4,15 +4,7 @@ import { useQuery } from "@tanstack/react-query";
 import Loading from "../../Components/Loading/Loading";
 
 const AllSellers = () => {
-  // const [sellers, setSellers]= useState([])
-  //    useEffect(
-  //     ()=>{
-  //         axios.get('http://localhost:5000/sellers')
-  //         .then(data=> console.log(data.data.data))
-  //     }, []
-  //    )
-
-  const url = "http://localhost:5000/sellers";
+  const url = "https://laptopserver.vercel.app/sellers";
   const {
     data: sellers = [],
     isLoading,
@@ -28,15 +20,19 @@ const AllSellers = () => {
   });
 
   const deleteSeller = (seller) => {
-    
-    fetch(`http://localhost:5000/sellers/${seller._id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        console.log(data);
-        refetch();
-      });
+    const proceed = window.confirm(
+      "Are you sure you want to delete this user?"
+    );
+    if (proceed) {
+      fetch(`https://laptopserver.vercel.app/sellers/${seller._id}`, {
+        method: "DELETE",
+      })
+        .then((res) => res.json())
+        .then((data) => {
+          console.log(data);
+          refetch();
+        });
+    }
   };
 
   if (isLoading) {
@@ -70,7 +66,9 @@ const AllSellers = () => {
               </td>
               <td>
                 <button
-                onClick={()=>{deleteSeller(seller)}}
+                  onClick={() => {
+                    deleteSeller(seller);
+                  }}
                   type="button"
                   class="text-white bg-gradient-to-r from-red-400 via-red-500 to-red-600 hover:bg-gradient-to-br focus:ring-4 focus:outline-none focus:ring-red-300 dark:focus:ring-red-800 font-medium rounded-lg text-sm px-5 py-2.5 text-center mr-2 mb-2"
                 >
